Everything seems to be compiling.

+++ src/pfs/drive/drive.go
@@ -18,20 +18,20 @@ type ReaderAtCloser interface {
 // Driver represents a low-level pfs storage driver.
 type Driver interface {
 	RepoCreate(repo *pfs.Repo) error
-	RepoInspect(repo *pfs.Repo, shard int) (*pfs.RepoInfo, error)
-	RepoList(shard int) ([]*pfs.RepoInfo, error)
-	RepoDelete(repo *pfs.Repo, shard map[int]bool) error
-	CommitStart(parent *pfs.Commit, commit *pfs.Commit, shard map[int]bool) (*pfs.Commit, error)
-	CommitFinish(commit *pfs.Commit, shard map[int]bool) error
-	CommitInspect(commit *pfs.Commit, shard int) (*pfs.CommitInfo, error)
-	CommitList(repo *pfs.Repo, shard int) ([]*pfs.CommitInfo, error)
-	CommitDelete(commit *pfs.Commit, shard map[int]bool) error
-	FilePut(file *pfs.File, shard int, offset int64, reader io.Reader) error
-	MakeDirectory(file *pfs.File, shard map[int]bool) error
-	FileGet(file *pfs.File, shard int) (ReaderAtCloser, error)
-	FileInspect(file *pfs.File, shard int) (*pfs.FileInfo, error)
-	FileList(file *pfs.File, shard int) ([]*pfs.FileInfo, error)
-	FileDelete(file *pfs.File, shard int) error
-	DiffPull(commit *pfs.Commit, shard int, diff io.Writer) error
+	RepoInspect(repo *pfs.Repo, shard uint64) (*pfs.RepoInfo, error)
+	RepoList(shard uint64) ([]*pfs.RepoInfo, error)
+	RepoDelete(repo *pfs.Repo, shard map[uint64]bool) error
+	CommitStart(parent *pfs.Commit, commit *pfs.Commit, shard map[uint64]bool) (*pfs.Commit, error)
+	CommitFinish(commit *pfs.Commit, shard map[uint64]bool) error
+	CommitInspect(commit *pfs.Commit, shard uint64) (*pfs.CommitInfo, error)
+	CommitList(repo *pfs.Repo, shard uint64) ([]*pfs.CommitInfo, error)
+	CommitDelete(commit *pfs.Commit, shard map[uint64]bool) error
+	FilePut(file *pfs.File, shard uint64, offset int64, reader io.Reader) error
+	MakeDirectory(file *pfs.File, shard map[uint64]bool) error
+	FileGet(file *pfs.File, shard uint64) (ReaderAtCloser, error)
+	FileInspect(file *pfs.File, shard uint64) (*pfs.FileInfo, error)
+	FileList(file *pfs.File, shard uint64) ([]*pfs.FileInfo, error)
+	FileDelete(file *pfs.File, shard uint64) error
+	DiffPull(commit *pfs.Commit, shard uint64, diff io.Writer) error
 	DiffPush(commit *pfs.Commit, diff io.Reader) error
 }
